# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Liveness probe payload returned by ``GET /health``.",
        "properties": {
          "service": {
            "description": "Service name reported by the running process.",
            "examples": [
              "family-office-portal"
            ],
            "title": "Service",
            "type": "string"
          },
          "status": {
            "description": "Service status indicator.",
            "examples": [
              "ok"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "service"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "RefreshLogEntry": {
        "description": "One row of the refresh history for a backend service.",
        "properties": {
          "is_stale": {
            "description": "True when the cached dataset has exceeded its staleness window.",
            "examples": [
              false
            ],
            "title": "Is Stale",
            "type": "boolean"
          },
          "last_error_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 timestamp of the most recent failed refresh, if any.",
            "examples": [
              "2026-05-15T22:14:01Z"
            ],
            "title": "Last Error At"
          },
          "last_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Truncated error message from the most recent failure.",
            "examples": [
              "upstream returned HTTP 500"
            ],
            "title": "Last Error Message"
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 timestamp of the most recent successful refresh.",
            "examples": [
              "2026-05-16T03:00:00Z"
            ],
            "title": "Last Success At"
          },
          "service": {
            "description": "Backend service identifier.",
            "examples": [
              "llc-manager"
            ],
            "title": "Service",
            "type": "string"
          }
        },
        "required": [
          "service",
          "is_stale"
        ],
        "title": "RefreshLogEntry",
        "type": "object"
      },
      "RefreshStatusResponse": {
        "description": "Response payload for ``GET /admin/refresh-status``.",
        "properties": {
          "entries": {
            "description": "One entry per backend service tracked by the scheduler.",
            "items": {
              "$ref": "#/components/schemas/RefreshLogEntry"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "title": "RefreshStatusResponse",
        "type": "object"
      },
      "RefreshTriggerRequest": {
        "description": "Request body for ``POST /admin/refresh/{service}``.",
        "properties": {
          "force": {
            "default": false,
            "description": "When true, bypass the staleness check and refresh immediately even if cached data is fresh.",
            "examples": [
              true
            ],
            "title": "Force",
            "type": "boolean"
          }
        },
        "title": "RefreshTriggerRequest",
        "type": "object"
      },
      "RefreshTriggerResponse": {
        "description": "Response payload for ``POST /admin/refresh/{service}``.",
        "properties": {
          "forced": {
            "description": "True when the staleness check was bypassed.",
            "examples": [
              false
            ],
            "title": "Forced",
            "type": "boolean"
          },
          "scheduled": {
            "description": "True when the refresh job was enqueued.",
            "examples": [
              true
            ],
            "title": "Scheduled",
            "type": "boolean"
          },
          "service": {
            "description": "Backend service whose refresh job was scheduled.",
            "examples": [
              "llc-manager"
            ],
            "title": "Service",
            "type": "string"
          }
        },
        "required": [
          "service",
          "scheduled",
          "forced"
        ],
        "title": "RefreshTriggerResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Byron Williams"
    },
    "description": "Private read-only family estate portal aggregating entity, document, finance, and portfolio data from internal backend services behind a SQLite read-through cache. Identity comes from Authentik forward auth (ADR-004). Page routes return server-rendered HTML; admin and health routes return JSON. See docs/planning/tech-spec.md for the contract.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Family Office Portal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "description": "Render the landing dashboard: totals, upcoming dates, recent documents.\n\nAuthentication: Viewer or Admin (ADR-004).\n\nArgs:\n    request (Request): Current request.\n\nReturns:\n    Response: Rendered dashboard page.",
        "operationId": "home__get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Home dashboard",
        "tags": [
          "home"
        ]
      }
    },
    "/admin/refresh-status": {
      "get": {
        "description": "Return the latest success and failure per backend service.\n\nAuthentication: Admin only (ADR-004). Services that have never run are\nomitted.\n\nReturns:\n    RefreshStatusResponse: One entry per service that has run.",
        "operationId": "refresh_status_admin_refresh_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Per-service refresh log",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/refresh/{service}": {
      "post": {
        "description": "Run the named refresh job after the response is sent.\n\nAuthentication: Admin only (ADR-004).\n\nArgs:\n    service (Literal[\"entities\", \"holdings\", \"positions\", \"documents\"]):\n        Dataset to refresh.\n    body (RefreshTriggerRequest): Trigger options. Refreshes always run\n        now; ``force`` is echoed for compatibility.\n    background (BackgroundTasks): FastAPI background task queue.\n\nReturns:\n    RefreshTriggerResponse: Confirmation that the job was queued.",
        "operationId": "trigger_refresh_admin_refresh__service__post",
        "parameters": [
          {
            "in": "path",
            "name": "service",
            "required": true,
            "schema": {
              "enum": [
                "entities",
                "holdings",
                "positions",
                "documents"
              ],
              "title": "Service",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTriggerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "summary": "Trigger a manual refresh",
        "tags": [
          "admin"
        ]
      }
    },
    "/documents": {
      "get": {
        "description": "Render documents grouped by category, with optional name search.\n\nWorks without JavaScript: the search form submits here with ``q``.\nAuthentication: Viewer or Admin (ADR-004). Viewers never see\nconfidential documents (D-14).\n\nArgs:\n    request (Request): Current request.\n    q (OptionalSearch): Optional name search text.\n\nReturns:\n    Response: Rendered documents page.",
        "operationId": "documents_index_documents_get",
        "parameters": [
          {
            "description": "Optional name search.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional name search.",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Document folders",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/search": {
      "get": {
        "description": "Return the search results fragment used by the Documents page.\n\nArgs:\n    request (Request): Current request.\n    q (RequiredSearch): Text to find in document names.\n\nReturns:\n    Response: HTML fragment listing matching documents.",
        "operationId": "documents_search_documents_search_get",
        "parameters": [
          {
            "description": "Text to find in names.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Text to find in names.",
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search documents by name (HTMX partial)",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{document_id}/download": {
      "get": {
        "description": "Download a document.\n\nThe file proxy to llc-manager lands with MVP task C5; until then a known\ndocument returns 503.\n\nArgs:\n    request (Request): Current request.\n    document_id (str): Document identifier.\n\nRaises:\n    HTTPException: 404 when unknown or not visible, 503 until C5.",
        "operationId": "document_download_documents__document_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Document not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Not yet available"
          }
        },
        "summary": "Download a document",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{document_id}/preview": {
      "get": {
        "description": "Show a document inline.\n\nThe file proxy to llc-manager lands with MVP task C5; until then a known\ndocument returns 503.\n\nArgs:\n    request (Request): Current request.\n    document_id (str): Document identifier.\n\nRaises:\n    HTTPException: 404 when unknown or not visible, 503 until C5.",
        "operationId": "document_preview_documents__document_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Document not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Not yet available"
          }
        },
        "summary": "Inline document preview",
        "tags": [
          "documents"
        ]
      }
    },
    "/entities": {
      "get": {
        "description": "Render every cached LLC and trust.\n\nAuthentication: Viewer or Admin (ADR-004).\n\nArgs:\n    request (Request): Current request.\n\nReturns:\n    Response: Rendered entity list.",
        "operationId": "entities_index_entities_get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Entity list",
        "tags": [
          "entities"
        ]
      }
    },
    "/entities/{entity_id}": {
      "get": {
        "description": "Render one entity with its documents.\n\nAuthentication: Viewer or Admin (ADR-004).\n\nArgs:\n    request (Request): Current request.\n    entity_id (str): Entity identifier.\n\nReturns:\n    Response: Rendered entity page.\n\nRaises:\n    HTTPException: 404 when the entity is not cached.",
        "operationId": "entity_detail_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entity detail",
        "tags": [
          "entities"
        ]
      }
    },
    "/finances": {
      "get": {
        "description": "Render account totals and crypto positions from the cache.\n\nAuthentication: Viewer or Admin (ADR-004).\n\nArgs:\n    request (Request): Current request.\n\nReturns:\n    Response: Rendered finances page.",
        "operationId": "finances_finances_get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Account totals and digital currency",
        "tags": [
          "finances"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Return a static liveness payload.\n\nUnauthenticated. Intended for orchestrator probes and the Newman CI\nworkflow so polling can begin before any backend services are reachable.\n\nReturns:\n    HealthResponse: Status indicator and service name.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "health"
        ]
      }
    },
    "/portfolio": {
      "get": {
        "description": "Render cached holdings from pp-security-master.\n\nAuthentication: Viewer or Admin (ADR-004).\n\nArgs:\n    request (Request): Current request.\n\nReturns:\n    Response: Rendered portfolio page.",
        "operationId": "portfolio_portfolio_get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Holdings",
        "tags": [
          "portfolio"
        ]
      }
    }
  },
  "tags": [
    {
      "description": "Liveness probes for orchestrators.",
      "name": "health"
    },
    {
      "description": "Landing dashboard.",
      "name": "home"
    },
    {
      "description": "Document folders, search, and previews.",
      "name": "documents"
    },
    {
      "description": "Account totals and digital currency.",
      "name": "finances"
    },
    {
      "description": "Holdings from pp-security-master.",
      "name": "portfolio"
    },
    {
      "description": "LLC and trust views from llc-manager.",
      "name": "entities"
    },
    {
      "description": "Refresh status and manual refresh triggers (Admin role).",
      "name": "admin"
    }
  ]
}